pyproject.toml
[project]
name = "qogita"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.5",
    "fastapi[standard]>=0.115.12",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "missingno>=0.5.2",
    "nltk>=3.9.1",
    "numpy==1.26",
    "pandas>=2.2.3",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
    "xgboost>=3.0.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "mlflow>=2.21.3",
]
